# src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Mock Server",
    "version" : "1.0"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
    "/api/correspondents" : {
      "get" : {
        "tags" : [ "Correspondents" ],
        "operationId" : "GetCorrespondents",
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        }
      },
      "post" : {
        "tags" : [ "Correspondents" ],
        "operationId" : "CreateCorrespondent",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NewCorrespondent"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NewCorrespondent"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NewCorrespondent"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        }
      }
    },
    "/api/correspondents/{id}" : {
      "put" : {
        "tags" : [ "Correspondents" ],
        "operationId" : "UpdateCorrespondent",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Correspondent"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Correspondent"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Correspondent"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        }
      },
      "delete" : {
        "tags" : [ "Correspondents" ],
        "operationId" : "DeleteCorrespondent",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        }
      }
    },
    "/api/documents" : {
      "get" : {
        "tags" : [ "Documents" ],
        "operationId" : "GetDocuments",
        "parameters" : [ {
          "name" : "Page",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "page_size",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "query",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "ordering",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "tags__id__all",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "integer",
              "format" : "int32"
            }
          }
        }, {
          "name" : "document_type__id",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "correspondent__id",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "truncate_content",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        }
      }
    },
    "/api/documents/{id}" : {
      "put" : {
        "tags" : [ "Documents" ],
        "operationId" : "UpdateDocument",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Document"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Document"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Document"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        }
      },
      "delete" : {
        "tags" : [ "Documents" ],
        "operationId" : "DeleteDocument",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        }
      }
    },
...
    "/api/documents/{id}/thumb" : {
      "get" : {
        "tags" : [ "Documents" ],
        "operationId" : "GetDocumentThumb",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        }
      }
    },
    "/api/documents/{id}/download" : {
      "get" : {
        "tags" : [ "Documents" ],
        "operationId" : "DownloadDocument",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "original",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        }
      }
    },
    "/api/documents/{id}/preview" : {
      "get" : {
        "tags" : [ "Documents" ],
        "operationId" : "GetDocumentPreview",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        }
      }
    },
    "/api/documents/{id}/metadata" : {
      "get" : {
        "tags" : [ "Documents" ],
        "operationId" : "GetDocumentMetadata",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        }
      }
    },
    "/api/document_types" : {
      "get" : {
        "tags" : [ "DocumentTypes" ],
        "operationId" : "GetDocumentTypes",
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        }
      },
      "post" : {
        "tags" : [ "DocumentTypes" ],
        "operationId" : "CreateDocumentType",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NewDocumentType"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NewDocumentType"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NewDocumentType"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        }
      }
    },
    "/api/document_types/{id}" : {
      "put" : {
        "tags" : [ "DocumentTypes" ],
        "operationId" : "UpdateDocumentType",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DocumentType"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DocumentType"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DocumentType"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        }
      },
      "delete" : {
        "tags" : [ "DocumentTypes" ],
        "operationId" : "DeleteDocumentType",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "Correspondent" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "slug" : {
            "type" : "string",
            "nullable" : true
          },
          "name" : {
            "type" : "string",
            "nullable" : true
          },
          "match" : {
            "type" : "string",
            "nullable" : true
          },
          "matching_algorithm" : {
            "type" : "integer",
            "format" : "int64"
          },
          "is_insensitive" : {
            "type" : "boolean"
          },
          "document_count" : {
            "type" : "integer",
            "format" : "int64"
          },
          "last_correspondence" : {
            "type" : "string",
            "format" : "date-time"
          }
        },
        "additionalProperties" : false
      },
      "Document" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int32"
          },
          "correspondent" : {
            "type" : "integer",
            "format" : "int32",
            "nullable" : true
          },
          "document_type" : {
            "type" : "integer",
            "format" : "int32",
            "nullable" : true
          },
          "storage_path" : {
            "type" : "integer",
            "format" : "int32",
            "nullable" : true
          },
          "title" : {
            "type" : "string",
            "nullable" : true
          },
          "content" : {
            "type" : "string",
            "nullable" : true
          },
          "tags" : {
            "type" : "array",
            "nullable" : true,
            "items" : {
              "type" : "integer",
              "format" : "int32"
            }
          },
          "created" : {
            "type" : "string",
            "format" : "date-time"
          },
          "created_date" : {
            "type" : "string",
            "format" : "date-time"
          },
          "modified" : {
            "type" : "string",
            "format" : "date-time"
          },
          "added" : {
            "type" : "string",
            "format" : "date-time"
          },
          "archive_serial_number" : {
            "type" : "string",
            "nullable" : true
          },
          "original_file_name" : {
            "type" : "string",
            "nullable" : true
          },
          "archived_file_name" : {
            "type" : "string",
            "nullable" : true
          }
        },
        "additionalProperties" : false
      },
      "DocumentType" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "slug" : {
            "type" : "string",
            "nullable" : true
          },
          "name" : {
            "type" : "string",
            "nullable" : true
          },
          "match" : {
            "type" : "string",
            "nullable" : true
          },
          "matching_algorithm" : {
            "type" : "integer",
            "format" : "int64"
          },
          "is_insensitive" : {
            "type" : "boolean"
          },
          "document_count" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "additionalProperties" : false
      },
      "NewCorrespondent" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "nullable" : true
          },
          "match" : {
            "type" : "string",
            "nullable" : true
          },
          "matching_algorithm" : {
            "type" : "integer",
            "format" : "int64"
          },
          "is_insensitive" : {
            "type" : "boolean"
          },
          "document_count" : {
            "type" : "integer",
            "format" : "int64"
          },
          "last_correspondence" : {
            "type" : "string",
            "format" : "date-time"
          }
        },
        "additionalProperties" : false
      },
      "NewDocumentType" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "nullable" : true
          },
          "match" : {
            "type" : "string",
            "nullable" : true
          },
          "matching_algorithm" : {
            "type" : "integer",
            "format" : "int64"
          },
          "is_insensitive" : {
            "type" : "boolean"
          }
        },
        "additionalProperties" : false
      },
...
    }
  }
}